feat: Add Products Page hook
